requirements.txt
alembic==1.17.1
blinker==1.9.0
click==8.3.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
